Increased the wait to make sure the test passes in PR builds.
