+++ apps/tests/ui/page/modal-page.ts
@@ -2,7 +2,7 @@
 import TKUnit = require("../../TKUnit");
 
 export function onShownModally(args: ShownModallyData) {
-    TKUnit.wait(0.100);
+    TKUnit.wait(0.350);
     args.context.shownModally = true;
     args.closeCallback("return value");
 }
